Split into distict files.
This will come in handy when we want to test the code. This is also the
official advice:
<https://publish.obsidian.md/hub/04+-+Guides%2C+Workflows%2C+%26+Courses/Guides/How+to+test+plugin+code+that+uses+Obsidian+APIs>
